Refactor some modules
