+++ .pre-commit-config.yaml
@@ -18,8 +18,6 @@ repos:
     rev: 5.6.3
     hooks:
       - id: isort
-        args:
-        - "--line-length=88"
         name: isort (python)
 
   # flake8
