+++ setup.cfg
